Add universal AppDrawer and integrate across screens
