+++ lib/features/training/training_screen.dart
@@ -3,7 +3,7 @@ import 'package:provider/provider.dart';
 
 import 'package:bugbear_app/features/training/notifier/session_notifier.dart';
 import 'package:bugbear_app/features/training/widgets/training_header.dart';
-import 'package:bugbear_app/features/training/widgets/training_drawer.dart';
+import 'package:bugbear_app/widgets/app_drawer.dart';
 import 'package:bugbear_app/features/training/widgets/progress_row.dart';
 import 'package:bugbear_app/features/training/widgets/exercise_canvas.dart';
 import 'package:bugbear_app/features/training/widgets/control_button_row.dart';
@@ -20,7 +20,7 @@ class TrainingScreen extends StatelessWidget {
     final current = exercises[idx];
 
     return Scaffold(
-      drawer: const TrainingDrawer(), // neuer Drawer
+      drawer: const AppDrawer(),
       appBar: TrainingHeader(
         phaseName: 'Phase ${state.phaseId}',
         onHelpPressed: () {
